Corrección de guardar la vida del personaje.

+++ raylibproject/src/cuchau/Personaje.h
@@ -117,7 +117,7 @@ public:
     bool        get_isPlayer()   const { return isPlayer; }              // Es controlado por teclado?
     //void        set_isPlayer(bool ip)  { isPlayer = ip; }
     std::string_view GetNombre()      const { return Player.nombre; }         // Nombre del personaje
-	int 	 GetRangoMax()     const { return Player.rango_max; }      // Rango maximo del ataque (en unidades del juego, para calcular alcance de proyectiles)
+	float 	 GetRangoMax()     const { return Player.rango_max; }      // Rango maximo del ataque (en unidades del juego, para calcular alcance de proyectiles)
     //float       GetCharSize()    const { return Size3D; }              // Tamano del billboard
     //Texture2D*  GetAtaqueTexture()     { return &Ataque; }               // Textura del proyectil
     //float       GetAttackSpeed() const { return Player.attack_speed * SPEED_SCALE; } // Vel. proyectil en u/s
